+++ src/API/Dashboard/EphemeralUploads/Id/Edit/Get/Route.hs
@@ -1,0 +1,24 @@
+module API.Dashboard.EphemeralUploads.Id.Edit.Get.Route where
+
+--------------------------------------------------------------------------------
+
+import Data.Text (Text)
+import Domain.Types.Cookie (Cookie)
+import Domain.Types.HxRequest (HxRequest)
+import Effects.Database.Tables.EphemeralUploads qualified as EphemeralUploads
+import Lucid qualified
+import Servant ((:>))
+import Servant qualified
+import Text.HTML (HTML)
+
+--------------------------------------------------------------------------------
+
+-- | "GET /dashboard/ephemeral-uploads/:ephemeral_upload_id/edit"
+type Route =
+  "dashboard"
+    :> "ephemeral-uploads"
+    :> Servant.Capture "ephemeral_upload_id" EphemeralUploads.Id
+    :> "edit"
+    :> Servant.Header "Cookie" Cookie
+    :> Servant.Header "HX-Request" HxRequest
+    :> Servant.Get '[HTML] (Servant.Headers '[Servant.Header "HX-Redirect" Text] (Lucid.Html ()))